requirements.txt
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
pandas>=2.1.0,<3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
nltk>=3.8.0,<4.0.0
python-multipart>=0.0.6
boto3>=1.34.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
bcrypt>=4.0.0